Refactor: Service 계층에서 로직 실행되도록 변경
