+++ src/main/java/hongik/discordbots/initializer/service/PythonFileService.java
@@ -3,10 +3,10 @@ package hongik.discordbots.initializer.service;
 import hongik.discordbots.initializer.dto.FileDownloadResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.io.ByteArrayResource;
-import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 
 import java.io.ByteArrayOutputStream;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -18,14 +18,12 @@ import java.util.zip.ZipOutputStream;
 @RequiredArgsConstructor
 public class PythonFileService {
 
-    private final FileResourceService fileResourceService;
-
     // 실행파일 + 봇파일 등 포함해서 zip으로 만들어주는 매서드
     public FileDownloadResponse createPythonBotZip(List<String> dependencies) {
         try {
 
             // 여기서 실질적인 파이썬 봇 파일 생성
-            ByteArrayResource mainPyResource = fileResourceService.createCombinedPythonFile(dependencies);
+            ByteArrayResource mainPyResource = createPythonBotFile(dependencies);
 
             ByteArrayOutputStream baos = new ByteArrayOutputStream();
             try (ZipOutputStream zos = new ZipOutputStream(baos)) {
@@ -51,5 +49,52 @@ public class PythonFileService {
         } catch (Exception e) {
             throw new RuntimeException("Error creating Python environment ZIP", e);
         }
+    }
+
+    // 원하는 기능을 가진 파이썬 봇 파일 생성하는 매서드
+    private ByteArrayResource createPythonBotFile(List<String> dependencies) {
+        // Load the contents of each Python file
+        String header = loadFileContent("src/main/resources/static/Python/header.py");
+        String footer = loadFileContent("src/main/resources/static/Python/footer.py");
+
+        // Combine the contents based on the dependencies
+        StringBuilder mainPyContent = new StringBuilder();
+        mainPyContent.append(header).append("\n");
+
+        // 이 부분은 추후에 바꿔야 할 듯 이름을 매개변수로 받는 매서드 생성해서
+        if (dependencies.contains("pingpong")) {
+            mainPyContent.append(loadFileContent("src/main/resources/static/Python/pingpong.py")).append("\n");
+        }
+
+        if (dependencies.contains("say")) {
+            mainPyContent.append(loadFileContent("src/main/resources/static/Python/say.py")).append("\n");
+        }
+
+        mainPyContent.append(footer);
+
+        return new ByteArrayResource(mainPyContent.toString().getBytes());
+    }
+
+    // 봇 파일 내부에 추가할 내용 불러오는 매서드
+    private String loadFileContent(String filePath) {
+        try {
+            Path path = Paths.get(filePath).toAbsolutePath().normalize();
+            if (Files.exists(path)) {
+                return Files.readString(path);
+            } else {
+                throw new IllegalArgumentException("File not found: " + filePath);
+            }
+        } catch (IOException e) {
+            throw new RuntimeException("Error reading file: " + filePath, e);
+        }
+    }
+
+
+    // 이 매서드는 아직 실행되지 않음
+    public FileDownloadResponse createDependenciesZip(String programmingLanguage, List<String> dependencies) {
+        // Create a zip file containing the dependencies for the specified programming language
+        // This is a placeholder implementation
+        byte[] data = "".getBytes();
+        return new FileDownloadResponse(programmingLanguage + "-dependencies.zip", "application/zip", data);
     }
 }
